requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
tqdm==4.65.0
pydantic==1.10.7